requirements.txt
tqdm==4.65.0
google-generativeai
streamlit==1.40.0
selenium==4.25.0
requests==2.28.1